refactor: improve file structure by splitting script.js
