+++ public/library.html
@@ -60,5 +60,5 @@
 </div>
 
 </body>
-<script src="./js/script.js"></script>
+<script type="module" src="./js/main.js"></script>
 </html>
